feat: Move Summary functionality to SummaryRepo
* Split CR Repo (large) into two smaller reports with summary logic go into its own repo
* Update tests
