+++ backend/lcfs/web/api/compliance_report/repo.py
@@ -1,5 +1,3 @@
-from collections import defaultdict
-
 import asyncio
 import structlog
 from datetime import datetime
@@ -10,7 +8,6 @@ from sqlalchemy import (
     and_,
     asc,
     desc,
-    update,
     String,
     cast,
     or_,
@@ -18,9 +15,8 @@ from sqlalchemy import (
     exists,
 )
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.inspection import inspect
 from sqlalchemy.orm import joinedload
-from typing import List, Optional, Dict, Union, TypedDict, Type
+from typing import List, Optional, TypedDict, Type
 
 from lcfs.db.dependencies import get_async_db_session
 from lcfs.db.models.comment import ComplianceReportInternalComment
@@ -46,12 +42,7 @@ from lcfs.db.models.compliance.FuelExport import FuelExport
 from lcfs.db.models.compliance.FuelSupply import FuelSupply
 from lcfs.db.models.compliance.NotionalTransfer import NotionalTransfer
 from lcfs.db.models.compliance.OtherUses import OtherUses
-from lcfs.db.models.fuel.ExpectedUseType import ExpectedUseType
-from lcfs.db.models.fuel.FuelCategory import FuelCategory
-from lcfs.db.models.fuel.FuelType import FuelType
-from lcfs.db.models.initiative_agreement.InitiativeAgreement import InitiativeAgreement
 from lcfs.db.models.organization.Organization import Organization
-from lcfs.db.models.transfer.Transfer import Transfer
 from lcfs.db.models.user.Role import RoleEnum
 from lcfs.db.models.user.UserProfile import UserProfile
 from lcfs.web.api.base import (
@@ -62,19 +53,10 @@ from lcfs.web.api.base import (
 from lcfs.web.api.compliance_report.schema import (
     ComplianceReportBaseSchema,
     ComplianceReportViewSchema,
-    ComplianceReportSummaryUpdateSchema,
 )
 from lcfs.web.api.fuel_supply.repo import FuelSupplyRepository
 from lcfs.web.api.role.schema import user_has_roles
 from lcfs.web.core.decorators import repo_handler
-
-from lcfs.web.api.compliance_report.dtos import (
-    ChangelogFuelSuppliesDTO,
-    ChangelogOtherUsesDTO,
-    ChangelogAllocationAgreementsDTO,
-    ChangelogFuelExportsDTO,
-    ChangelogNotionalTransfersDTO,
-)
 
 logger = structlog.get_logger(__name__)
 
@@ -529,28 +511,6 @@ class ComplianceReportRepository:
         ]
 
     @repo_handler
-    async def get_fuel_type(self, fuel_type_id: int) -> FuelType:
-        return await self.db.scalar(
-            select(FuelType).where(FuelType.fuel_type_id == fuel_type_id)
-        )
-
-    @repo_handler
-    async def get_fuel_category(self, fuel_category_id: int) -> FuelCategory:
-        return await self.db.scalar(
-            select(FuelCategory).where(
-                FuelCategory.fuel_category_id == fuel_category_id
-            )
-        )
-
-    @repo_handler
-    async def get_expected_use(self, expected_use_type_id: int) -> ExpectedUseType:
-        return await self.db.scalar(
-            select(ExpectedUseType).where(
-                ExpectedUseType.expected_use_type_id == expected_use_type_id
-            )
-        )
-
-    @repo_handler
     async def update_compliance_report(
         self, report: ComplianceReport
     ) -> ComplianceReportBaseSchema:
@@ -589,260 +549,6 @@ class ComplianceReportRepository:
             await self.db.rollback()
             logger.error(f"Error updating compliance report: {e}")
             raise
-
-    @repo_handler
-    async def add_compliance_report_summary(
-        self, summary: ComplianceReportSummary
-    ) -> ComplianceReportSummary:
-        """
-        Adds a new compliance report summary to the database.
-        """
-        self.db.add(summary)
-        await self.db.flush()
-        await self.db.refresh(summary)
-        return summary
-
-    @repo_handler
-    async def reset_summary_lock(self, compliance_report_id: int):
-        query = (
-            update(ComplianceReportSummary)
-            .where(ComplianceReportSummary.compliance_report_id == compliance_report_id)
-            .values(is_locked=False)
-        )
-        await self.db.execute(query)
-        return True
-
-    @repo_handler
-    async def save_compliance_report_summary(
-        self, summary: ComplianceReportSummaryUpdateSchema
-    ):
-        """
-        Save the compliance report summary to the database.
-
-        :param summary: The generated summary data
-        """
-        existing_summary = await self.get_summary_by_report_id(
-            summary.compliance_report_id
-        )
-
-        if existing_summary:
-            summary_obj = existing_summary
-        else:
-            raise ValueError(
-                f"""No summary found with report ID {
-                    summary.compliance_report_id}"""
-            )
-
-        summary_obj.is_locked = summary.is_locked
-        # Update renewable fuel target summary
-        for row in summary.renewable_fuel_target_summary:
-            line_number = row.line
-            for fuel_type in ["gasoline", "diesel", "jet_fuel"]:
-                column_name = f"""line_{line_number}_{
-                    row.field.lower()}_{fuel_type}"""
-                setattr(summary_obj, column_name, int(getattr(row, fuel_type)))
-
-        # Update low carbon fuel target summary
-        for row in summary.low_carbon_fuel_target_summary:
-            column_name = f"line_{row.line}_{row.field}"
-            setattr(
-                summary_obj,
-                column_name,
-                int(row.value),
-            )
-
-        # Update non-compliance penalty summary
-        non_compliance_summary = summary.non_compliance_penalty_summary
-        for row in non_compliance_summary:
-            if row.line == 11:
-                summary_obj.line_11_fossil_derived_base_fuel_total = row.total_value
-            elif row.line == 21:
-                summary_obj.line_21_non_compliance_penalty_payable = row.total_value
-            elif row.line is None:  # Total row
-                summary_obj.total_non_compliance_penalty_payable = row.total_value
-
-        self.db.add(summary_obj)
-        await self.db.flush()
-        await self.db.refresh(summary_obj)
-        return summary_obj
-
-    @repo_handler
-    async def get_summary_by_report_id(self, report_id: int) -> ComplianceReportSummary:
-        query = select(ComplianceReportSummary).where(
-            ComplianceReportSummary.compliance_report_id == report_id
-        )
-
-        result = await self.db.execute(query)
-        return result.scalars().first()
-
-    # @repo_handler
-    # async def get_summary_versions(self, report_id: int):
-    #     query = (
-    #         select(
-    #             ComplianceReportSummary.summary_id,
-    #             ComplianceReportSummary.version,
-    #             case(
-    #                 (
-    #                     ComplianceReportSummary.supplemental_report_id.is_(None),
-    #                     "Original",
-    #                 ),
-    #                 else_="Supplemental",
-    #             ).label("type"),
-    #         )
-    #         .where(ComplianceReportSummary.compliance_report_id == report_id)
-    #         .order_by(ComplianceReportSummary.version)
-    #     )
-
-    #     result = await self.db.execute(query)
-    #     return result.all()
-
-    @repo_handler
-    async def get_transferred_out_compliance_units(
-        self,
-        compliance_period_start: datetime,
-        compliance_period_end: datetime,
-        organization_id: int,
-    ) -> int:
-        result = await self.db.scalar(
-            select(func.sum(Transfer.quantity)).where(
-                Transfer.agreement_date.between(
-                    compliance_period_start, compliance_period_end
-                ),
-                Transfer.from_organization_id == organization_id,
-                Transfer.current_status_id == 6,  # Recorded
-            )
-        )
-        return result or 0
-
-    @repo_handler
-    async def get_received_compliance_units(
-        self,
-        compliance_period_start: datetime,
-        compliance_period_end: datetime,
-        organization_id: int,
-    ) -> int:
-        result = await self.db.scalar(
-            select(func.sum(Transfer.quantity)).where(
-                Transfer.agreement_date.between(
-                    compliance_period_start, compliance_period_end
-                ),
-                Transfer.to_organization_id == organization_id,
-                Transfer.current_status_id == 6,  # Recorded
-            )
-        )
-        return result or 0
-
-    @repo_handler
-    async def get_issued_compliance_units(
-        self,
-        compliance_period_start: datetime,
-        compliance_period_end: datetime,
-        organization_id: int,
-    ) -> int:
-        result = await self.db.scalar(
-            select(func.sum(InitiativeAgreement.compliance_units)).where(
-                InitiativeAgreement.transaction_effective_date.between(
-                    compliance_period_start, compliance_period_end
-                ),
-                InitiativeAgreement.to_organization_id == organization_id,
-                InitiativeAgreement.current_status_id == 3,  # Approved
-            )
-        )
-        return result or 0
-
-    def aggregate_quantities(
-        self, records: List[Union[FuelSupply, OtherUses]], fossil_derived: bool
-    ) -> Dict[str, float]:
-        """Common aggregation logic for both FuelSupply and OtherUses"""
-        fuel_quantities = defaultdict(float)
-
-        for record in records:
-            # Check if record matches fossil_derived filter
-            if (
-                isinstance(record, FuelSupply)
-                and record.fuel_type.fossil_derived == fossil_derived
-            ):
-                fuel_category = self._format_category(record.fuel_category.category)
-
-                total_quantity = (
-                    (record.quantity or 0)
-                    + (record.q1_quantity or 0)
-                    + (record.q2_quantity or 0)
-                    + (record.q3_quantity or 0)
-                    + (record.q4_quantity or 0)
-                )
-                fuel_quantities[fuel_category] += total_quantity
-            elif (
-                isinstance(record, OtherUses)
-                and record.fuel_type.fossil_derived == fossil_derived
-            ):
-                fuel_category = self._format_category(record.fuel_category.category)
-                fuel_quantities[fuel_category] += record.quantity_supplied
-
-        return dict(fuel_quantities)
-
-    @repo_handler
-    async def aggregate_other_uses_quantity(
-        self, compliance_report_id: int, fossil_derived: bool
-    ) -> Dict[str, float]:
-        """Aggregate quantities from other uses."""
-        query = (
-            select(
-                FuelCategory.category,
-                func.coalesce(func.sum(OtherUses.quantity_supplied), 0).label(
-                    "quantity"
-                ),
-            )
-            .select_from(OtherUses)
-            .join(FuelType, OtherUses.fuel_type_id == FuelType.fuel_type_id)
-            .join(
-                FuelCategory,
-                OtherUses.fuel_category_id == FuelCategory.fuel_category_id,
-            )
-            .where(
-                OtherUses.compliance_report_id == compliance_report_id,
-                FuelType.fossil_derived.is_(fossil_derived),
-                FuelType.other_uses_fossil_derived.is_(fossil_derived),
-            )
-            .group_by(FuelCategory.category)
-        )
-
-        result = await self.db.execute(query)
-        return {self._format_category(row.category): row.quantity for row in result}
-
-    @repo_handler
-    async def aggregate_allocation_agreements(
-        self, compliance_report_id: int
-    ) -> Dict[str, float]:
-        """Aggregate quantities from allocation agreements for renewable fuels."""
-        query = (
-            select(
-                FuelCategory.category,
-                func.coalesce(func.sum(AllocationAgreement.quantity), 0).label(
-                    "quantity"
-                ),
-            )
-            .select_from(AllocationAgreement)
-            .join(FuelType, AllocationAgreement.fuel_type_id == FuelType.fuel_type_id)
-            .join(
-                FuelCategory,
-                AllocationAgreement.fuel_category_id == FuelCategory.fuel_category_id,
-            )
-            .where(
-                AllocationAgreement.compliance_report_id == compliance_report_id,
-                FuelType.fossil_derived.is_(False),
-                FuelType.other_uses_fossil_derived.is_(False),
-            )
-            .group_by(FuelCategory.category)
-        )
-
-        result = await self.db.execute(query)
-        return {self._format_category(row.category): row.quantity for row in result}
-
-    @staticmethod
-    def _format_category(category: str) -> str:
-        """Format the fuel category string."""
-        return category.lower().replace(" ", "_")
 
     @repo_handler
     async def get_all_org_reported_years(self, organization_id: int):
@@ -913,35 +619,6 @@ class ComplianceReportRepository:
             .where(ComplianceReport.legacy_id == legacy_id)
         )
         return result.scalars().unique().first()
-
-    @repo_handler
-    async def get_compliance_report_group_id(self, report_id):
-        """
-        Retrieve the compliance report group ID
-        """
-        result = await self.db.scalar(
-            select(ComplianceReport.compliance_report_group_uuid).where(
-                ComplianceReport.compliance_report_id == report_id
-            )
-        )
-        return result
-
-    async def get_previous_summary(
-        self, compliance_report: ComplianceReport
-    ) -> ComplianceReportSummary:
-        result = await self.db.execute(
-            select(ComplianceReport)
-            .options(
-                joinedload(ComplianceReport.summary),
-            )
-            .where(
-                ComplianceReport.compliance_report_group_uuid
-                == compliance_report.compliance_report_group_uuid,
-                ComplianceReport.version == compliance_report.version - 1,
-            )
-            .limit(1)
-        )
-        return result.scalars().first().summary
 
     @repo_handler
     async def delete_compliance_report(self, compliance_report_id: int) -> bool:
