+++ backend/lcfs/web/api/compliance_report/summary_service.py
@@ -27,6 +27,9 @@ from lcfs.web.api.compliance_report.schema import (
     ComplianceReportSummarySchema,
     ComplianceReportSummaryUpdateSchema,
 )
+from lcfs.web.api.compliance_report.summary_repo import (
+    ComplianceReportSummaryRepository,
+)
 from lcfs.web.api.fuel_export.repo import FuelExportRepository
 from lcfs.web.api.fuel_supply.repo import FuelSupplyRepository
 from lcfs.web.api.notional_transfer.services import NotionalTransferServices
@@ -76,7 +79,8 @@ def get_compliance_data_service():
 class ComplianceReportSummaryService:
     def __init__(
         self,
-        repo: ComplianceReportRepository = Depends(),
+        repo: ComplianceReportSummaryRepository = Depends(),
+        cr_repo: ComplianceReportRepository = Depends(),
         trxn_repo: TransactionRepository = Depends(),
         notional_transfer_service: NotionalTransferServices = Depends(
             NotionalTransferServices
@@ -92,6 +96,7 @@ class ComplianceReportSummaryService:
         ),
     ):
         self.repo = repo
+        self.cr_repo = cr_repo
         self.notional_transfer_service = notional_transfer_service
         self.trxn_repo = trxn_repo
         self.fuel_supply_repo = fuel_supply_repo
@@ -397,7 +402,7 @@ class ComplianceReportSummaryService:
     ) -> ComplianceReportSummarySchema:
         """Several fields on Report Summary are Transient until locked, this function will re-calculate fields as necessary"""
         # Fetch the compliance report details
-        compliance_report = await self.repo.get_compliance_report_by_id(
+        compliance_report = await self.cr_repo.get_compliance_report_by_id(
             report_id, is_model=True
         )
         if not compliance_report:
@@ -406,7 +411,7 @@ class ComplianceReportSummaryService:
         prev_compliance_report = None
         if not compliance_report.supplemental_initiator:
             prev_compliance_report = (
-                await self.repo.get_assessed_compliance_report_by_period(
+                await self.cr_repo.get_assessed_compliance_report_by_period(
                     compliance_report.organization_id,
                     int(compliance_report.compliance_period.description) - 1,
                 )
